src/requirements.txt
google-generativeai>=0.3.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.7
python-dotenv>=1.0.0

